MULTIMODAL/requirements.txt
# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Computer Vision
Pillow>=9.0.0
opencv-python>=4.7.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress bars and logging
tqdm>=4.64.0
wandb>=0.15.0

# Configuration and utilities
pyyaml>=6.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# Image processing and augmentation
albumentations>=1.3.0
timm>=0.9.0

# Additional utilities
einops>=0.6.0
packaging>=21.0

# Development tools (optional)
jupyter>=1.0.0
ipywidgets>=8.0.0
black>=23.0.0
flake8>=5.0.0

# For VMamba and Mamba dependencies
mamba-ssm>=1.0.0
causal-conv1d>=1.0.0
triton>=2.0.0

# For efficient training
accelerate>=0.20.0
deepspeed>=0.9.0

# Memory optimization
psutil>=5.9.0
py3nvml>=0.2.7 